Use DWORD instead of unsigned long for TSS key type to be consistent with the MSDN docs for TlsAlloc.

+++ asio/include/asio/detail/win_tss_ptr.hpp
@@ -72,7 +72,7 @@ public:
 private:
   // Thread-specific storage to allow unlocked access to determine whether a
   // thread is a member of the pool.
-  unsigned long tss_key_;
+  DWORD tss_key_;
 };
 
 } // namespace detail
